Letterboxing detection, v2 sector settings
Set up counters and checks to set the flag for the letterboxing mode. This will probably take some tweaking to get right, but it's a start.

Add data settings and ui handlers for switching between v1 and v2 sector mapping. V2 sector mapping gives us double the number of sectors to utilize around the border...which means better color fidelity and granularity.