Add right/bottom LED counts, optimize splitter, cleanup...
Add right and bottom counts for LEDs in addition to top and left, because that's nicer.
Optimize our splitter to only do what it needs to do, instead of split all the things all the time.

Make WLEDs use a subset of the "full" colors, versus splitting individually for each device. Much faster.