Better brightness control, crop timer, auto-disable timeout
Fix brightness controls for several devices so we're using the actual device's brightness setting, versus the more memory-intensive way of clamping colors.

Fix averaging and timer for cropping check. Instead of trying to get some kind of average for a region, we're individually checking the color of each pixel in a row/column, and "failing" if that value is higher than a certain level.

Fix autodisable timeout not actually timing out.