Use same branch on update, fix video capture
Now that we're recording the branch from Github directly in the app, we can easily check what branch we're using, and then pass that branch to the update script(s) so that we aren't switching back and forth.

Also, fix method to instantiate video capture device, as emgu.cv's method has changed.